Change arch to all.
Almost all of the code is arch-independent. With more and more
non-mipsel receivers appearing, it's no longer feasible to just
say "mipsel".

Only dfplib.so is mipsel arch. We'll take the chance. Apparently
it's used for "pearl photo frames", which might be used on vu+duo2,
which does have a mipsel arch. It's definitely not used on the
vu+solo4k which has arm arch.